breakdown the classification module into algorithms and wrappers
