+++ pyriemann_qiskit/__init__.py
@@ -1,9 +1,8 @@
-from . import autoencoders, classification, ensemble, pipelines
+from . import autoencoders, ensemble, pipelines
 from ._version import __version__
 
 __all__ = [
     "__version__",
-    "classification",
     "pipelines",
     "ensemble",
     "autoencoders",
