+++ pyriemann_qiskit/classification/__init__.py
@@ -1,0 +1,10 @@
+from .algorithms import NearestConvexHull
+from .wrappers import QuanticMDM, QuanticNCH, QuanticSVM, QuanticVQC
+
+__all__ = [
+    "NearestConvexHull",
+    "QuanticMDM",
+    "QuanticNCH",
+    "QuanticSVM",
+    "QuanticVQC"
+]
